move server files to src/ and add a test for appMiddleware
