+++ scripts/serveUniversal.js
@@ -1,0 +1,2 @@
+import startServer from '../src/server'
+startServer()
